# Conflicts:
#	STM32F1/libraries/SPI/src/SPI.cpp
#	STM32F4/cores/maple/libmaple/HardwareSerial.cpp
#	STM32F4/cores/maple/libmaple/adc.h
#	STM32F4/cores/maple/libmaple/dmaF4.c
#	STM32F4/cores/maple/libmaple/dmaF4.h
#	STM32F4/cores/maple/libmaple/gpio.c
#	STM32F4/cores/maple/libmaple/gpio.h
#	STM32F4/cores/maple/libmaple/rcc.c
#	STM32F4/cores/maple/wirish_digital.cpp
#	STM32F4/cores/maple/wirish_time.h
#	STM32F4/libraries/SPI/src/SPI.cpp
#	STM32F4/platform.txt
#	STM32F4/system/libmaple/Arduino.h
#	STM32F4/variants/generic_f407v/generic_f407v.h
#	STM32F4/variants/generic_f407v/ld/common.inc
#	STM32F4/variants/generic_f407v/ld/jtag.ld
#	STM32F4/variants/generic_f407v/pin_map.c
#	STM32F4/variants/generic_f407v/variant.h
#	STM32F4/variants/generic_f407v/wirish/start.S

+++ STM32F4/cores/maple/io.h
@@ -121,7 +121,7 @@ void pinMode(uint8 pin, WiringPinMode mode);
  * @param value Either LOW (write a 0) or HIGH (write a 1).
  * @see pinMode()
  */
-void digitalWrite(uint8 pin, uint16 value);
+void digitalWrite(uint8 pin, uint8 value);
 
 /**
  * Read a digital value from a pin.  The pin must have its mode set to
